requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymongo==4.10.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.2
sse-starlette==2.1.3
python-dotenv==1.0.1
pydantic[email]==2.9.2
